requirements.txt
vsts-cli==0.1.4
requests==2.21.0
qhue==1.0.12